APIs/evetech.net/0.8.6/python-aiohttp-client/tests/test_alliance_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.bad_request import BadRequest
from openapi_server.models.error_limited import ErrorLimited
from openapi_server.models.gateway_timeout import GatewayTimeout
from openapi_server.models.get_alliances_alliance_id_icons_not_found import GetAlliancesAllianceIdIconsNotFound
from openapi_server.models.get_alliances_alliance_id_icons_ok import GetAlliancesAllianceIdIconsOk
from openapi_server.models.get_alliances_alliance_id_not_found import GetAlliancesAllianceIdNotFound
from openapi_server.models.get_alliances_alliance_id_ok import GetAlliancesAllianceIdOk
from openapi_server.models.internal_server_error import InternalServerError
from openapi_server.models.service_unavailable import ServiceUnavailable


pytestmark = pytest.mark.asyncio

async def test_get_alliances(client):
    """Test case for get_alliances

    List all alliances
    """
    params = [('datasource', tranquility)]
    headers = { 
        'Accept': 'application/json',
        'if_none_match': 'if_none_match_example',
    }
    response = await client.request(
        method='GET',
        path='/latest/alliances/',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_alliances_alliance_id(client):
    """Test case for get_alliances_alliance_id

    Get alliance information
    """
    params = [('datasource', tranquility)]
    headers = { 
        'Accept': 'application/json',
        'if_none_match': 'if_none_match_example',
    }
    response = await client.request(
        method='GET',
        path='/latest/alliances/{alliance_id}'.format(alliance_id=56),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_alliances_alliance_id_corporations(client):
    """Test case for get_alliances_alliance_id_corporations

    List alliance's corporations
    """
    params = [('datasource', tranquility)]
    headers = { 
        'Accept': 'application/json',
        'if_none_match': 'if_none_match_example',
    }
    response = await client.request(
        method='GET',
        path='/latest/alliances/{alliance_id}/corporations'.format(alliance_id=56),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_alliances_alliance_id_icons(client):
    """Test case for get_alliances_alliance_id_icons

    Get alliance icon
    """
    params = [('datasource', tranquility)]
    headers = { 
        'Accept': 'application/json',
        'if_none_match': 'if_none_match_example',
    }
    response = await client.request(
        method='GET',
        path='/latest/alliances/{alliance_id}/icons'.format(alliance_id=56),
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

